refactor(main): extract account, cache administration and progress sync
Three more concerns move out of main.lua as controllers taking the plugin
instance, following the lib/qr_login.lua host pattern:

- lib/account.lua: the requireLogin gate every online feature calls, plus
  account status, cookie renewal and clearing credentials.
- lib/cache_admin.lua: the download directory (validate, pick, migrate
  books to it), the cache overview with per-book sizes and cleanup, and
  importing content already on disk via lib/scan.lua. The two safety
  rules that make deletion safe on a user-selected library folder are
  documented at the top of the module.
- lib/progress_sync.lua: the unfinished progress-sync experiment. Only
  the upload path has a live entry point (the weread_sync_progress
  Dispatcher action); the pull and reader-URL parsing halves had no
  caller in main.lua and are kept here so the whole WIP feature sits in
  one place.

Bodies were moved verbatim; only call paths changed. i18n keys unchanged.

main.lua: 2869 -> 2253 lines.
